+++ pkg/get/helm_test.go
@@ -2,7 +2,9 @@
 // Licensed under the Apache License, Version 2.0. See LICENSE file in the project root for full license information.
 package get
 
-import "testing"
+import (
+	"testing"
+)
 
 func TestDownloadHelm(t *testing.T) {
 	if err := DownloadHelm(); err != nil {