+++ pkg/get/k3d.go
@@ -2,7 +2,9 @@
 // Licensed under the Apache License, Version 2.0. See LICENSE file in the project root for full license information.
 package get
 
-import "fmt"
+import (
+	"fmt"
+)
 
 const k3dVersion = "v1.7.0"